Rename wasm C sources to avoid conflicts with standard wasm headers
