+++ lib/src/lib.c
@@ -11,4 +11,4 @@
 #include "./subtree.c"
 #include "./tree_cursor.c"
 #include "./tree.c"
-#include "./wasm.c"
+#include "./wasm_store.c"
