+++ lib/src/wasm_store.h
